+++ start_remote.sh
@@ -41,9 +41,16 @@ while getopts 'd:s:p:g:j:' OPTION; do
 done
 
 # Check if required argument is provided
-if [[ -z $DEVELOPER_NAME ]]
+if [ -z $DEVELOPER_NAME ]
 then
     echo "Missing argument -d name"
+    help
+fi
+
+# Check if project folder exists
+if [ ! -d "${SOURCE_CODE_DIR}/${PROJECT}/" ]
+then
+    printf "Project folder ${SOURCE_CODE_DIR}/${PROJECT}/ does not exist, please check the arguments \n"
     help
 fi
 
@@ -81,8 +88,13 @@ rm -rf .display_${DISPLAY_NUMBER}
 mkdir -p .display_${DISPLAY_NUMBER}/socket
 touch .display_${DISPLAY_NUMBER}/Xauthority
 
-# Copy project requirements.txt to context
-cp ${SOURCE_CODE_DIR}/${PROJECT}/requirements.txt ./requirements.txt
+# Copy project requirements.txt to context, if it does not exist, create an empty requirements.txt
+if [ -f "${SOURCE_CODE_DIR}/${PROJECT}/requirements.txt" ]
+then
+    cp ${SOURCE_CODE_DIR}/${PROJECT}/requirements.txt ./requirements.txt
+else
+    touch ./requirements.txt
+fi
 
 # Build the image
 docker build \
@@ -112,7 +124,7 @@ socat -d -d -d TCP4:localhost:60${DISPLAY_NUMBER} UNIX-LISTEN:.display_${DISPLAY
 NUM_CAM=$(ls -dl /dev/video* | grep '^c' | wc -l)
 
 # Launch the container
-if [ $NUM_CAM = 2 ]
+if [ $NUM_CAM -ge 2 ]
 then
     docker run -it --rm \
         --name ${ML_CONTAINER_NAME} \
